refactor: move sign in logic to App component
