+++ src/components/App/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { UserSession, AppConfig } from 'blockstack';
-import { configure } from 'radiks';
+import { configure, getConfig, User } from 'radiks';
 
 import Login from '../Login';
 import UploadFile from '../UploadFile';
@@ -13,6 +13,16 @@ configure({
   apiServer: process.env.RADIKS_SERVER || 'http://localhost:1260',
   userSession
 });
+
+const handleSignIn = async () => {
+  const { userSession } = getConfig();
+  if (userSession.isSignInPending()) {
+    await userSession.handlePendingSignIn();
+    await User.createWithCurrentUser();
+  }  
+};
+
+handleSignIn();
 
 const App = () => {
   return (
