+++ src/components/UploadFile/index.js
@@ -4,7 +4,6 @@ import Button from '@material-ui/core/Button';
 import Fab from '@material-ui/core/Fab';
 import AddIcon from '@material-ui/icons/Add';
 import { makeStyles } from '@material-ui/core/styles';
-import { User, getConfig } from 'radiks';
 import SelectUser from '../SelectUser';
 
 const useStyles = makeStyles(theme => ({
@@ -17,16 +16,7 @@ const useStyles = makeStyles(theme => ({
   })
 );
 
-const handleSignIn = async () => {
-  const { userSession } = getConfig();
-  if (userSession.isSignInPending()) {
-    await userSession.handlePendingSignIn();
-    await User.createWithCurrentUser();
-  }  
-};
-
 const UploadFile = () => {
-  handleSignIn();
   const classes = useStyles();
   return (
     <div>
